Macro vfront_basics::diag_types

macro_rules! diag_types {
    ($component : ident, $($kind : ident $typ : ident = $desc : literal), * $(,)
 ?) => { ... };
}

Defines a list of diagnostic types. To be used once in every component that defines diagnostics.

This macro creates two things:

For every diagnostic type, a pub const is defined of type &'static DiagType and the same name as the type, which can then be used as argument to DiagSystem::begin to emit a diagnostic.

A pub fn called register_diags is defined, taking one argument of type &mut DiagRegistry, which will register all diagnostics defined here.

Takes a component name (identifier, preferably same as the crate name), followed by comma-separated diagnostic type definitions. Every diagnostic type definition consists of the diagnostic kind (one of the keywords: note_ignored, note, warning_ignored, warning, error, fatal), the diagnostic name (an identifier), an equals sign, and the description (a string literal).

Example:

use vfront_basics::diag::{DiagSystem, DiagRegistry, DiagStore};

mod diags {
    #[macro_use]
    use vfront_basics::diag_types;
    diag_types!(example_component,
        error redefined_var = "A variable has been defined more than once",
        warning unused_var = "A variable has been defined, but has never been used",
    );
}

// ...

fn main() {
    let mut registry = DiagRegistry::new();
    diags::register_diags(&mut registry);
    let sink = DiagStore::new();
    let diags = DiagSystem::new(&registry, &sink);

    // ...

    diags.begin(diags::unused_var, "variable foo is unused")
        .emit();
}