Module vfront_basics::source

A bunch of data structures that keep the source code in memory and handle storage and processing of source location data. The idea is similar to clang’s SourceManager:

There is one SourceManager for the whole design that contains all source code involved.

The source code is made of SourceChunks, which are either files or macro expansions, and are wholesale loaded into memory as a single str.

All SourceChunks are conceptually concatenated into one big linear array of bytes, and every source location can thus be stored as a single u32 that is a byte index into this virtual array (called SourceLoc). Note that this array is never actually materialized — we just store the virtual start position of every SourceChunk and use binary search when looking up a SourceLoc.

For immediate access to the source, we also have SourceRef, which can be considered an unpacked form of SourceLoc. This can be used to directly access the backing storage of a given SourceChunk.

SourceRange is a range of two SourceLoc that belong to the same chunk, SourceRangeRef is an unpacked version, which also can be directly used as as string.

When it is necessary to print a diagnostic, run SourceManager::get_line_info on a location to obtain its decoded source position. If the decoded position is within a macro, use SourceChunkInfo::MacroExpansion::loc_invoked to obtain its definition place, and keep printing position data from there. If the decoded position is an included file, use SourceChunkInfo::File::loc_included to iterate up the include stack.

When a simple file+line+column triple is needed (for export to something that has a fixed idea of a location), SourceManager::get_simple_line_info can be called to obtain such. This will automatically go up the macro invocation stack, if any, to obtain an actual file location.

To support Verilog `line directive and similar constructs, line override entries can be added to chunks. Doing this will cause SourceManager::get_line_info to return the overriden location data in addition to the true location. Note that line override entries must be added in position order.

Structs

SourceChunk

A contiguous chunk of source code, belonging to a SourceManager. Can be a source file or a macro expansion.

SourceLineInfo

Result of looking up a line number.

SourceLoc

A compressed representation of a location in the source code, to be used in ASTs and other long-term storage places. Only has meaning in relation to a given SourceManager, and can be unpacked into a SourceRef when necessary. Consists of a single NonZeroU32 (so that Option is as cheap as a plain SourceLoc, when necessary).

SourceManager

A catalogue of all source files that were involved in the creation of an AST / a design / whatever. New files can be inserted into this given only a shared reference.

SourceOverrideLineInfo

Information about overriden line number and file name, from `line.

SourceRange

A half-open range of two SourceLocs that must belong to the same SourceChunk.

SourceRangeRef

An uncompressed representation of a range of locations in the source code, analogous to SourceRef. Can be dereferenced to obtain underlying str.

SourceRef

An uncompressed representation of a location in the source code, to be used for currently-processed code when space is not at premium. The chunk_pos is a direct index in chunk.text and can be used to retrieve the text.

SourceSimpleLineInfo

Simple decoded location, to be used when exporting to formats that need plain file+line+column triple. This always gives a file location — when macros are involved, this is the location of the top macro invocation.

Enums

SourceChunkInfo

Information about how a SourceChunk was created.

SourceLineOverrideKind

Describes the kind of a line override entry (the third argument of `line).