Struct vfront_basics::source::SourceReader

pub struct SourceReader<'a> { /* private fields */ }

A helper for reading from a source chunk character by character.

This maintains three pointers into a chunk:

The cursor: points at the next character to be read.

The end: points at a “stop point” past which the cursor cannot advance.

The bookmark: points at an arbitrary “start point” at or before the cursor, used for rollback and range construction. Initially points at the same place as the cursor.

A reader can be constructed (by Reader::from) from a SourceRef (sets the cursor to given position, sets the end to the end of the chunk) or a SourceRangeRef (sets the cursor to the start of the range, sets the end to the end of the range).

Implementations

impl<'a> SourceReader<'a>

pub fn suffix(&self) -> &'a str

Returns the chunk slice between the cursor and the end.

pub fn move_to(&mut self, loc: SourceRef<'a>)

Moves the cursor to the given position.

pub fn cursor(&self) -> SourceRef<'a>

Returns the current cursor position.

pub fn end(&self) -> SourceRef<'a>

Returns the end position of this reader.

pub fn bookmark(&self) -> SourceRef<'a>

Returns the current bookmark position.

pub fn set_mark(&mut self)

Moves the bookmark position to the current cursor position.

pub fn advance(&mut self, num: usize)

Moves the cursor position forwards by the given number of bytes.

pub fn rollback(&mut self)

Moves the cursor position back to the current bookmark position (effectively undoing all eat calls since the last SourceReader::set_mark call.

pub fn range(&self) -> SourceRangeRef<'a>

Returns a SourceRangeRef starting from the current bookmark position and ending at the current cursor position.

pub fn range_from(&self, start: SourceRef<'a>) -> SourceRangeRef<'a>

Returns a SourceRangeRef spanning from the given position to the current cursor position.

pub fn peek(&self) -> (Option<char>, SourceRef<'a>)

Returns the character at the cursor, or None if the cursor is at the end position. Also returns a SourceRef that can be passed to SourceReader::move_to to consume the returned character.

pub fn eat(&mut self) -> Option<char>

Returns the character at the cursor and consumes it, or returns None if the cursor is at the end position.

pub fn eat_if<P>(&mut self, pred: P) -> Option<char> where
P: FnOnce(char) -> bool,

Checks if the character at the cursor satisfies the given predicate and, if so, consumes and returns it. If the cursor is at the end position or the character doesn’t satisfy the predicate, returns None and doesn’t consume anything.

pub fn eat_if_map<F, T>(&mut self, f: F) -> Option<T> where
F: FnOnce(char) -> Option<T>,

Takes the character at the cursor, runs a given function returning an Option on it, returns its result. If the function returns a value other than None, consumes the character.

pub fn eat_while<P>(&mut self, pred: P) -> &'a str where
P: FnMut(char) -> bool,

Consumes the characters at the cursor, stopping at the end position or when a character is found that doesn’t satisfy the given predicate. Returns a string slice of all consumed characters.

pub fn try_eat(&mut self, s: &str) -> bool

If there is an occurence of the given string directly at the cursor, consumes it and returns true. Otherwise, does nothing and returns false.

pub fn is_empty(&self) -> bool

Returns true iff the cursor position is at the end position.