Module vfront_basics::diag

The vfront diagnostic system.

The system has several major moving parts:

The main concept in the system is the Diagnostic, which is a message that we want to show to the user.

Every Diagnostic has some DiagType, which represents a particular kind of diagnostics.

DiagTypes have to be statically defined.

DiagRegistry keeps track of all DiagTypes defined in active components.

Every vfront component can define its own DiagTypes, and has to provide a function that registers its DiagTypes in the DiagRegistry. This is done by the diag_types macro.

Every Diagnostic has an associated DiagSeverity (ignore, note, warning, error, fatal), which determines its handling. The default severity of a diagnostic is determined by its DiagType, but can be overriden by the user on per-type basis.

A diagnostic also has:

The main message (a string).

Any number of DiagSpans.

An optional help message (which suggests a possible fix for the diagnostic).

A DiagSpan is a reference to some source code related to the diagnostic.

A DiagSpan has a SourceRange and a string message.

A DiagSpan can be primary (refers to the location of the problem), or secondary (refers to some code that is likely of interest, but is not the cause of the problem itself).

Every diagnostic SHOULD have at least one primary span. The only exception are problems that are really, truly not associated with any location in source code (think “undefined main function” in C).

The central point of diagnostic handling is DiagSystem — all code that can emit diagnostics should take a &DiagSystem as parameter, and a single DiagSystem is used for the whole processing run. Its responsibilities are:

Wrapping a DiagSink and DiagRegistry.

Storing per-type severity overrides and applying them.

Providing an easy interface to emit diagnostics.

Discarding ignored diagnostics.

Remembering whether an error-level diagnostic has been emitted (and thus the output should be considered broken, and processing aborted at some point).

Diagnostics are created and emitted using the DiagBuilder. The builder is obtained by calling DiagSystem::begin with a type and the main message, then various methods are called on it to add additional data, then finally it is emitted.

The final destination of diagnostics is pluggable, and represented by the DiagSink trait. The driver creates a DiagSink of whatever type it desires, then plugs it into the DiagSystem, which will notify the DiagSink of all emitted diagnostics.

For testing purposes, DiagStore is used as the diagnostics sink — it handles emitted diagnostics by simply storing them to an internal list, allowing for later comparison with “golden” diagnostics by the testcase code.

Structs

DiagBuilder

A builder for Diagnostics. Created by DiagSystem::begin. Once all options are set, DiagBuilder::emit should be called to actually emit the diagnostic.

DiagRegistry

A registry of all DiagTypes in the program. Should be constructed once at the beginning of the main function by calling the diagnostic registration function of every used component.

DiagSpan

A code span attached to a Diagnostic.

DiagStore

A DiagSink implementation that simply stores all emitted diagnostics in a vector. Can be converted to Vec<Diagnostic>.

DiagSystem

The main structure handling all diagnostics within the program. Generally created once per program run, but this is not a strict requirement.

DiagType

Describes a diagnostic type. Defined using diag_types.

Diagnostic

An emitted diagnostic.

Enums

DiagKind

The general kind of a DiagType. Determines its default severity and what severity levels it can be overriden to.

DiagSeverity

Determines the handling of a given diagnostic. Normally selected by the DiagType, but can be overriden by the user on per-type basis.

DiagSpanKind

The kind of a DiagSpan.

Traits

DiagSink

A consumer of Diagnostics.